time should not have been invented
script would fail to trigger if milliseconds val
ue was non-zero in the opportunity. I had previously handled up to seconds being non-zero, but my simulat
ion code didn't add jitter to milliseconds, hence the issue not being caught. This jitter is now added, and the functions that compute the opportunity to present a strategy are now flooring the timestamp before returning the opportunity timestamp